+++ CHANGES.rst
@@ -4,6 +4,10 @@ There's a frood who really knows where his towel is
 1.0b2 (unreleased)
 ------------------
 
+- Update view methods, sc.microsite content has the same methods than a folder. [jpgimenez]
+
+- Default view changed to folder_listing. [jpgimenez]
+
 - Update tests with behavior Navigation root. [cleberjsantos]
 
 - Apply behavior Navigation root. [agnogueira]